tools/api_fuzzer/requirements.txt
requests>=2.31.0
pyyaml>=6.0
graphql-core>=3.2.3
jsonschema>=4.19.0
openapi-spec-validator>=0.6.0
tqdm>=4.66.1
colorama>=0.4.6
aiohttp>=3.8.5
schemathesis>=3.19.0
click>=8.1.3
markupsafe>=2.1.1
jinja2>=3.0.0 